No need for parameters on discovery. Speed increase. Move some logic into Util. Move some from adapter to executor.
