+++ src/IKVM.JTReg.TestAdapter/IkvmJTRegTestAdapter.cs
@@ -1,15 +1,10 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices;
 using System.Security.Cryptography;
 using System.Text;
 
-using Microsoft.VisualStudio.TestPlatform.CoreUtilities.Extensions;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel;
-using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
 
 namespace IKVM.JTReg.TestAdapter
@@ -21,7 +16,7 @@ namespace IKVM.JTReg.TestAdapter
     public abstract class IkvmJTRegTestAdapter
     {
 
-        internal const int PARTITION_COUNT = 16;
+        internal const string URI = "executor://ikvmjtregtestadapter/v1";
 
         internal const string BASEDIR_PREFIX = "ikvm-jtreg-";
         internal const string TEST_ROOT_FILE_NAME = "TEST.ROOT";
@@ -34,41 +29,6 @@ namespace IKVM.JTReg.TestAdapter
         internal const string ENV_PREFIX = "JTREG_";
 
         protected static readonly MD5 MD5 = MD5.Create();
-
-        protected static readonly string[] DEFAULT_WINDOWS_ENV_VARS = { "PATH", "SystemDrive", "SystemRoot", "windir", "TMP", "TEMP", "TZ" };
-        protected static readonly string[] DEFAULT_UNIX_ENV_VARS = { "PATH", "DISPLAY", "GNOME_DESKTOP_SESSION_ID", "HOME", "LANG", "LC_ALL", "LC_CTYPE", "LPDEST", "PRINTER", "TZ", "XMODIFIERS" };
-
-        /// <summary>
-        /// Initializes the static instance.
-        /// </summary>
-        static IkvmJTRegTestAdapter()
-        {
-#if NETCOREAPP
-            // executable permissions may not have made it onto the JRE binaries so attempt to set them
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-            {
-                var javaHome = java.lang.System.getProperty("java.home");
-                foreach (var exec in new[] { "java", "javac", "jar", "jarsigner", "javadoc", "javah", "javap", "jdeps", "keytool", "policytool", "rmic", "wsgen", "wsimport" })
-                {
-                    var execPath = Path.Combine(javaHome, "bin", exec);
-                    if (File.Exists(execPath))
-                    {
-                        var psx = Mono.Unix.UnixFileSystemInfo.GetFileSystemEntry(execPath);
-                        var prm = psx.FileAccessPermissions;
-                        prm |= Mono.Unix.FileAccessPermissions.UserExecute;
-                        prm |= Mono.Unix.FileAccessPermissions.GroupExecute;
-                        prm |= Mono.Unix.FileAccessPermissions.OtherExecute;
-                        if (prm != psx.FileAccessPermissions)
-                            psx.FileAccessPermissions = prm;
-                    }
-                }
-            }
-#endif
-
-            // need to do some static configuration on the harness
-            if (JTRegTypes.Harness.GetClassDirMethod.invoke(null) == null)
-                JTRegTypes.Harness.SetClassDirMethod.invoke(null, JTRegTypes.ProductInfo.GetJavaTestClassDirMethod.invoke(null));
-        }
 
         /// <summary>
         /// Creates a short hash of the given string to uniquely identify it.
@@ -108,7 +68,6 @@ namespace IKVM.JTReg.TestAdapter
         /// <param name="logger"></param>
         /// <param name="baseDir"></param>
         /// <param name="output"></param>
-        /// <param name="errors"></param>
         /// <returns></returns>
         protected dynamic CreateTestManager(IMessageLogger logger, string baseDir, java.io.PrintWriter output)
         {
@@ -130,184 +89,6 @@ namespace IKVM.JTReg.TestAdapter
             Directory.CreateDirectory(reportDirectory);
             testManager.setReportDirectory(new java.io.File(reportDirectory));
             return testManager;
-        }
-
-        /// <summary>
-        /// Creates the parameters for a given suite.
-        /// </summary>
-        /// <param name="source"></param>
-        /// <param name="baseDir"></param>
-        /// <param name="testManager"></param>
-        /// <param name="testSuite"></param>
-        /// <param name="tests"></param>
-        /// <param name="debugUri"></param>
-        /// <returns></returns>
-        protected dynamic CreateParameters(string source, string baseDir, dynamic testManager, dynamic testSuite, IEnumerable<TestCase> tests, Uri debugUri)
-        {
-            if (source is null)
-                throw new ArgumentNullException(nameof(source));
-            if (baseDir is null)
-                throw new ArgumentNullException(nameof(baseDir));
-            if (testManager is null)
-                throw new ArgumentNullException(nameof(testManager));
-            if (testSuite is null)
-                throw new ArgumentNullException(nameof(testSuite));
-
-            // invoke new RegressionParameters(string, RegressionTestSuite)
-            var rp = JTRegTypes.RegressionParameters.New(DEFAULT_PARAM_TAG, testSuite);
-
-            // configure work directory
-            var wd = testManager.getWorkDirectory(testSuite);
-            rp.setWorkDirectory(wd);
-
-            var rd = testManager.getReportDirectory(testSuite);
-            rp.setReportDir(rd);
-
-            // if a ProblemList.txt or ExcludeList.txt file exists in the root, add them as exclude files
-            var excludeFileList = new List<java.io.File>();
-            foreach (var n in new[] { TEST_PROBLEM_LIST_FILE_NAME, TEST_EXCLUDE_LIST_FILE_NAME })
-                if (Path.Combine(((java.io.File)testSuite.getRootDir()).toString(), n) is string f && File.Exists(f))
-                    excludeFileList.Add(new java.io.File(new java.io.File(f).getAbsoluteFile().toURI().normalize()));
-
-            // if a IncludeList.txt file exists in the root, add it as include files
-            var includeFileList = new List<java.io.File>();
-            foreach (var n in new[] { TEST_INCLUDE_LIST_FILE_NAME })
-                if (Path.Combine(((java.io.File)testSuite.getRootDir()).toString(), n) is string f && File.Exists(f))
-                    includeFileList.Add(new java.io.File(new java.io.File(f).getAbsoluteFile().toURI().normalize()));
-
-            // passed in an explicit set of tests, add to an include file
-            if (tests != null)
-            {
-                // name of the current suite
-                var testSuiteName = Util.GetTestSuiteName(source, testSuite);
-
-                // fill in include list containing tests located within the suite
-                var includeListFile = (java.io.File)wd.getFile("IncludeList.txt");
-                using (var includeList = File.CreateText(includeListFile.toString()))
-                    foreach (var test in tests)
-                        if (string.Equals(test.GetPropertyValue(IkvmJTRegTestProperties.TestSuiteNameProperty), testSuiteName))
-                            includeList.WriteLine(test.GetPropertyValue(IkvmJTRegTestProperties.TestPathNameProperty) + " generic-all");
-
-                includeFileList.Add(new java.io.File(includeListFile.getAbsoluteFile().toURI().normalize()));
-            }
-
-            rp.setTests((java.util.Set)testManager.getTests(testSuite));
-            rp.setExecMode(testSuite.getDefaultExecMode() ?? JTRegTypes.ExecMode.AGENTVM);
-            rp.setCheck(false);
-            rp.setCompileJDK(JTRegTypes.JDK.Of(new java.io.File(java.lang.System.getProperty("java.home"))));
-            rp.setTestJDK(rp.getCompileJDK());
-            rp.setTestVMOptions(java.util.Collections.emptyList());
-            rp.setTestCompilerOptions(java.util.Collections.emptyList());
-            rp.setTestJavaOptions(java.util.Collections.emptyList());
-            rp.setFile((java.io.File)wd.getFile("config.jti"));
-            rp.setEnvVars(GetEnvVars(debugUri));
-            rp.setConcurrency(Environment.ProcessorCount);
-            rp.setTimeLimit(15000);
-            rp.setRetainArgs(java.util.Collections.singletonList("all"));
-            rp.setExcludeLists(excludeFileList.ToArray());
-            rp.setMatchLists(includeFileList.ToArray());
-            rp.setIgnoreKind(JTRegTypes.IgnoreKind.QUIET);
-            rp.setPriorStatusValues(null);
-            rp.setUseWindowsSubsystemForLinux(true);
-
-            // configure keywords to filter based on
-            string keywordsExpr = null;
-            keywordsExpr = CombineKeywords(keywordsExpr, "!ignore");
-            keywordsExpr = CombineKeywords(keywordsExpr, "!manual");
-            if (string.IsNullOrWhiteSpace(keywordsExpr) == false)
-                rp.setKeywordsExpr(keywordsExpr);
-
-            // locate and configure TestNG
-            var testNGPath = Path.Combine(Path.GetDirectoryName(typeof(IkvmJTRegTestAdapter).Assembly.Location), "jtreg", "testng.jar");
-            rp.setTestNGPath(JTRegTypes.SearchPath.New(testNGPath));
-
-            // locate and configure JUnit
-            var junitPath = Path.Combine(Path.GetDirectoryName(typeof(IkvmJTRegTestAdapter).Assembly.Location), "jtreg", "junit.jar");
-            rp.setJUnitPath(JTRegTypes.SearchPath.New(junitPath));
-
-            // locate and configure JUnit
-            var asmtoolsPath = Path.Combine(Path.GetDirectoryName(typeof(IkvmJTRegTestAdapter).Assembly.Location), "jtreg", "asmtools.jar");
-            rp.setAsmToolsPath(JTRegTypes.SearchPath.New(asmtoolsPath));
-
-            // final initialization
-            rp.initExprContext();
-            if ((bool)rp.isValid() == false)
-                throw new Exception();
-
-            return rp;
-        }
-
-        /// <summary>
-        /// Gets the set of environment variables to include with the tests by default.
-        /// </summary>
-        /// <param name="debugUri"></param>
-        /// <returns></returns>
-        java.util.Map GetEnvVars(Uri debugUri)
-        {
-            var envVars = new java.util.TreeMap();
-
-            // import existing variables based on the current OS
-            foreach (var var in ((string)JTRegTypes.OS.Current().family) == "windows" ? DEFAULT_WINDOWS_ENV_VARS : DEFAULT_UNIX_ENV_VARS)
-                if (Environment.GetEnvironmentVariable(var) is string val)
-                    envVars.put(var, val);
-
-            // import any variables prefixed with JTREG, as done by the existing tool
-            foreach (DictionaryEntry entry in Environment.GetEnvironmentVariables())
-                if (((string)entry.Key).StartsWith(ENV_PREFIX))
-                    envVars.put((string)entry.Key, (string)entry.Value);
-
-            // instruct child processes to signal debug attach
-            if (debugUri != null)
-            {
-                envVars.put("IKVM_DEBUG_URI", debugUri.ToString());
-                envVars.put("IKVM_DEBUG_WAIT", "60");
-            }
-
-            return envVars;
-        }
-
-        /// <summary>
-        /// Gets the results of an interview.
-        /// </summary>
-        /// <param name="parameters"></param>
-        /// <returns></returns>
-        /// <exception cref="ArgumentNullException"></exception>
-        protected IEnumerable<dynamic> GetTestResults(string source, dynamic testSuite, dynamic parameters)
-        {
-            if (source is null)
-                throw new ArgumentNullException(nameof(source));
-            if (testSuite is null)
-                throw new ArgumentNullException(nameof(testSuite));
-            if (parameters is null)
-                throw new ArgumentNullException(nameof(parameters));
-
-            // resolve RemainingToList generic method
-            var trtype = ikvm.runtime.Util.getInstanceTypeFromClass(JTRegTypes.TestResult.Class);
-            var rtltrmethod = typeof(java.util.IteratorExtensions).GetMethod(nameof(java.util.IteratorExtensions.RemainingToList)).MakeGenericMethod(trtype);
-
-            // wait until result is initialized
-            var trt = parameters.getWorkDirectory().getTestResultTable();
-            trt.waitUntilReady();
-
-            // find tests
-            var tests = parameters.getTests();
-            if (tests == null)
-                return ((IEnumerable<dynamic>)rtltrmethod.Invoke(null, new[] { (java.util.Iterator)trt.getIterator() })).OrderBy(i => Util.GetTestPathName(source, testSuite, i));
-            else if (tests.Length == 0)
-                return Array.Empty<object>();
-            else
-                return ((IEnumerable<dynamic>)rtltrmethod.Invoke(null, new[] { (java.util.Iterator)trt.getIterator() })).OrderBy(i => Util.GetTestPathName(source, testSuite, i));
-        }
-
-        /// <summary>
-        /// Combines the two keywords.
-        /// </summary>
-        /// <param name="kw1"></param>
-        /// <param name="kw2"></param>
-        /// <returns></returns>
-        string CombineKeywords(string kw1, string kw2)
-        {
-            return kw1 == null ? kw2 : kw1 + " & " + kw2;
         }
 
     }
