+++ src/IKVM.JTReg.TestAdapter/IkvmJTRegTestExecutor.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Runtime.InteropServices;
 using System.Threading;
 
 using Microsoft.VisualStudio.TestPlatform.ObjectModel;
@@ -13,9 +14,44 @@ using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
 namespace IKVM.JTReg.TestAdapter
 {
 
-    [ExtensionUri("executor://ikvmjtregtestadapter/v1")]
+    [ExtensionUri(URI)]
     public class IkvmJTRegTestExecutor : IkvmJTRegTestAdapter, ITestExecutor
     {
+
+        protected static readonly string[] DEFAULT_WINDOWS_ENV_VARS = { "PATH", "SystemDrive", "SystemRoot", "windir", "TMP", "TEMP", "TZ" };
+        protected static readonly string[] DEFAULT_UNIX_ENV_VARS = { "PATH", "DISPLAY", "GNOME_DESKTOP_SESSION_ID", "HOME", "LANG", "LC_ALL", "LC_CTYPE", "LPDEST", "PRINTER", "TZ", "XMODIFIERS" };
+
+        /// <summary>
+        /// Initializes the static instance.
+        /// </summary>
+        static IkvmJTRegTestExecutor()
+        {
+#if NETCOREAPP
+            // executable permissions may not have made it onto the JRE binaries so attempt to set them
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+            {
+                var javaHome = java.lang.System.getProperty("java.home");
+                foreach (var exec in new[] { "java", "javac", "jar", "jarsigner", "javadoc", "javah", "javap", "jdeps", "keytool", "policytool", "rmic", "wsgen", "wsimport" })
+                {
+                    var execPath = Path.Combine(javaHome, "bin", exec);
+                    if (File.Exists(execPath))
+                    {
+                        var psx = Mono.Unix.UnixFileSystemInfo.GetFileSystemEntry(execPath);
+                        var prm = psx.FileAccessPermissions;
+                        prm |= Mono.Unix.FileAccessPermissions.UserExecute;
+                        prm |= Mono.Unix.FileAccessPermissions.GroupExecute;
+                        prm |= Mono.Unix.FileAccessPermissions.OtherExecute;
+                        if (prm != psx.FileAccessPermissions)
+                            psx.FileAccessPermissions = prm;
+                    }
+                }
+            }
+#endif
+
+            // need to do some static configuration on the harness
+            if (JTRegTypes.Harness.GetClassDirMethod.invoke(null) == null)
+                JTRegTypes.Harness.SetClassDirMethod.invoke(null, JTRegTypes.ProductInfo.GetJavaTestClassDirMethod.invoke(null));
+        }
 
         CancellationTokenSource cts;
 
@@ -148,14 +184,12 @@ namespace IKVM.JTReg.TestAdapter
                 // we will need a full list of tests to apply any filters to
                 if (tests == null)
                 {
-                    var testCount = 0;
-
                     var l = new List<TestCase>();
 
                     // discover the full set of tests
                     foreach (dynamic testSuite in Util.GetTestSuites(source, testManager))
-                        foreach (var testResult in GetTestResults(source, testSuite, CreateParameters(source, baseDir, testManager, testSuite, null, debugUri)))
-                            l.Add(Util.ToTestCase(source, testSuite, testResult, testCount++ % PARTITION_COUNT));
+                        foreach (var testCase in Util.GetTestCases(source, testManager, testSuite))
+                            l.Add(testCase);
 
                     tests = l;
                 }
@@ -176,6 +210,152 @@ namespace IKVM.JTReg.TestAdapter
             {
                 frameworkHandle.SendMessage(TestMessageLevel.Error, $"JTReg: Exception occurred running tests for '{source}'.\n{e}");
             }
+        }
+
+        /// <summary>
+        /// Creates the parameters for a given suite.
+        /// </summary>
+        /// <param name="source"></param>
+        /// <param name="baseDir"></param>
+        /// <param name="testManager"></param>
+        /// <param name="testSuite"></param>
+        /// <param name="tests"></param>
+        /// <param name="debugUri"></param>
+        /// <returns></returns>
+        dynamic CreateParameters(string source, string baseDir, dynamic testManager, dynamic testSuite, IEnumerable<TestCase> tests, Uri debugUri)
+        {
+            if (source is null)
+                throw new ArgumentNullException(nameof(source));
+            if (baseDir is null)
+                throw new ArgumentNullException(nameof(baseDir));
+            if (testManager is null)
+                throw new ArgumentNullException(nameof(testManager));
+            if (testSuite is null)
+                throw new ArgumentNullException(nameof(testSuite));
+
+            // invoke new RegressionParameters(string, RegressionTestSuite)
+            var rp = JTRegTypes.RegressionParameters.New(DEFAULT_PARAM_TAG, testSuite);
+
+            // configure work directory
+            var wd = testManager.getWorkDirectory(testSuite);
+            rp.setWorkDirectory(wd);
+
+            // configure report directory
+            var rd = testManager.getReportDirectory(testSuite);
+            rp.setReportDir(rd);
+
+            // if a ProblemList.txt or ExcludeList.txt file exists in the root, add them as exclude files
+            var excludeFileList = new List<java.io.File>();
+            foreach (var n in new[] { TEST_PROBLEM_LIST_FILE_NAME, TEST_EXCLUDE_LIST_FILE_NAME })
+                if (Path.Combine(((java.io.File)testSuite.getRootDir()).toString(), n) is string f && File.Exists(f))
+                    excludeFileList.Add(new java.io.File(new java.io.File(f).getAbsoluteFile().toURI().normalize()));
+
+            // if a IncludeList.txt file exists in the root, add it as include files
+            var includeFileList = new List<java.io.File>();
+            foreach (var n in new[] { TEST_INCLUDE_LIST_FILE_NAME })
+                if (Path.Combine(((java.io.File)testSuite.getRootDir()).toString(), n) is string f && File.Exists(f))
+                    includeFileList.Add(new java.io.File(new java.io.File(f).getAbsoluteFile().toURI().normalize()));
+
+            // passed in an explicit set of tests, add to an include file
+            if (tests != null)
+            {
+                // name of the current suite
+                var testSuiteName = Util.GetTestSuiteName(source, testSuite);
+
+                // fill in include list containing tests located within the suite
+                var includeListFile = (java.io.File)wd.getFile("IncludeList.txt");
+                using (var includeList = File.CreateText(includeListFile.toString()))
+                    foreach (var test in tests)
+                        if (string.Equals(test.GetPropertyValue(IkvmJTRegTestProperties.TestSuiteNameProperty), testSuiteName))
+                            includeList.WriteLine(test.GetPropertyValue(IkvmJTRegTestProperties.TestPathNameProperty) + " generic-all");
+
+                includeFileList.Add(new java.io.File(includeListFile.getAbsoluteFile().toURI().normalize()));
+            }
+
+            rp.setTests((java.util.Set)testManager.getTests(testSuite));
+            rp.setExecMode(testSuite.getDefaultExecMode() ?? JTRegTypes.ExecMode.AGENTVM);
+            rp.setCheck(false);
+            rp.setCompileJDK(JTRegTypes.JDK.Of(new java.io.File(java.lang.System.getProperty("java.home"))));
+            rp.setTestJDK(rp.getCompileJDK());
+            rp.setTestVMOptions(java.util.Collections.emptyList());
+            rp.setTestCompilerOptions(java.util.Collections.emptyList());
+            rp.setTestJavaOptions(java.util.Collections.emptyList());
+            rp.setFile((java.io.File)wd.getFile("config.jti"));
+            rp.setEnvVars(GetEnvVars(debugUri));
+            rp.setConcurrency(Environment.ProcessorCount);
+            rp.setTimeLimit(15000);
+            rp.setRetainArgs(java.util.Collections.singletonList("all"));
+            rp.setExcludeLists(excludeFileList.ToArray());
+            rp.setMatchLists(includeFileList.ToArray());
+            rp.setIgnoreKind(JTRegTypes.IgnoreKind.QUIET);
+            rp.setPriorStatusValues(null);
+            rp.setUseWindowsSubsystemForLinux(true);
+
+            // configure keywords to filter based on
+            string keywordsExpr = null;
+            keywordsExpr = CombineKeywords(keywordsExpr, "!ignore");
+            keywordsExpr = CombineKeywords(keywordsExpr, "!manual");
+            if (string.IsNullOrWhiteSpace(keywordsExpr) == false)
+                rp.setKeywordsExpr(keywordsExpr);
+
+            // locate and configure TestNG
+            var testNGPath = Path.Combine(Path.GetDirectoryName(typeof(IkvmJTRegTestAdapter).Assembly.Location), "jtreg", "testng.jar");
+            rp.setTestNGPath(JTRegTypes.SearchPath.New(testNGPath));
+
+            // locate and configure JUnit
+            var junitPath = Path.Combine(Path.GetDirectoryName(typeof(IkvmJTRegTestAdapter).Assembly.Location), "jtreg", "junit.jar");
+            rp.setJUnitPath(JTRegTypes.SearchPath.New(junitPath));
+
+            // locate and configure JUnit
+            var asmtoolsPath = Path.Combine(Path.GetDirectoryName(typeof(IkvmJTRegTestAdapter).Assembly.Location), "jtreg", "asmtools.jar");
+            rp.setAsmToolsPath(JTRegTypes.SearchPath.New(asmtoolsPath));
+
+            // final initialization
+            rp.initExprContext();
+            if ((bool)rp.isValid() == false)
+                throw new Exception();
+
+            return rp;
+        }
+
+        /// <summary>
+        /// Gets the set of environment variables to include with the tests by default.
+        /// </summary>
+        /// <param name="debugUri"></param>
+        /// <returns></returns>
+        java.util.Map GetEnvVars(Uri debugUri)
+        {
+            var envVars = new java.util.TreeMap();
+
+            // import existing variables based on the current OS
+            foreach (var var in ((string)JTRegTypes.OS.Current().family) == "windows" ? DEFAULT_WINDOWS_ENV_VARS : DEFAULT_UNIX_ENV_VARS)
+                if (Environment.GetEnvironmentVariable(var) is string val)
+                    envVars.put(var, val);
+
+            // import any variables prefixed with JTREG, as done by the existing tool
+            foreach (DictionaryEntry entry in Environment.GetEnvironmentVariables())
+                if (((string)entry.Key).StartsWith(ENV_PREFIX))
+                    envVars.put((string)entry.Key, (string)entry.Value);
+
+            // instruct child processes to signal debug attach
+            if (debugUri != null)
+            {
+                envVars.put("IKVM_DEBUG_URI", debugUri.ToString());
+                envVars.put("IKVM_DEBUG_WAIT", "60");
+            }
+
+            return envVars;
+        }
+
+        /// <summary>
+        /// Combines the two keywords.
+        /// </summary>
+        /// <param name="kw1"></param>
+        /// <param name="kw2"></param>
+        /// <returns></returns>
+        string CombineKeywords(string kw1, string kw2)
+        {
+            return kw1 == null ? kw2 : kw1 + " & " + kw2;
         }
 
         /// <summary>
@@ -210,7 +390,7 @@ namespace IKVM.JTReg.TestAdapter
             cancellationToken.ThrowIfCancellationRequested();
 
             // only continue if there are in fact tests in the suite to execute
-            var firstTestResult = ((IEnumerable<object>)GetTestResults(source, testSuite, parameters)).FirstOrDefault();
+            var firstTestResult = ((IEnumerable<object>)Util.GetTestResults(source, testManager, testSuite)).FirstOrDefault();
             if (firstTestResult is null)
                 return;
